# spec/edited/openapi.yaml
openapi: 3.0.2
info:
  title: Wellbore DDMS OSDU
  description: build local
  version: '0.2'
paths:
...
  '/ddms/v2/dipsets/{dipsetid}/dips':
    get:
      tags:
        - Dips
      summary: Get dips
      description: |-
        Return dips from dipset from the given index until the given number of dips specifed in query parameters.
            If not specified returns all dips from dipset.
      operationId: get_dips
      parameters:
        - required: true
          schema:
            title: Dipsetid
            type: string
          name: dipsetid
          in: path
        - required: false
          schema:
            title: Index
            minimum: 0
            type: integer
          name: index
          in: query
        - required: false
          schema:
            title: Limit
            minimum: 0
            type: integer
          name: limit
          in: query
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Dips Ddms V2 Dipsets  Dipsetid  Dips Get
                type: array
                items:
                  $ref: '#/components/schemas/Dip'
        '404':
          description: DipSet not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
    post:
      tags:
        - Dips
      summary: Define the dips of the dipset
      description: Replace previous dips by provided dips. Sort dips by reference and azimuth.
      operationId: post_dips
      parameters:
        - description: The ID of the dipset
          required: true
          schema:
            title: Dipsetid
            type: string
            description: The ID of the dipset
          name: dipsetid
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      requestBody:
        content:
          application/json:
            schema:
              title: Dips
              type: array
              items:
                $ref: '#/components/schemas/Dip'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Post Dips Ddms V2 Dipsets  Dipsetid  Dips Post
                type: array
                items:
                  $ref: '#/components/schemas/Dip'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  '/ddms/v2/dipsets/{dipsetid}/dips/insert':
    post:
      tags:
        - Dips
      summary: insert dip in  a dipset
      description: |-
        Insert dips in dipset.
            Existing dips are not replaced.
            Several dip can have same reference.
            Operation will sort by reference all dips in dipset (may modify dip indexes).
      operationId: insert_dips
      parameters:
        - required: true
          schema:
            title: Dipsetid
            type: string
          name: dipsetid
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      requestBody:
        content:
          application/json:
            schema:
              title: Dips
              type: array
              items:
                $ref: '#/components/schemas/Dip'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Insert Dips Ddms V2 Dipsets  Dipsetid  Dips Insert Post
                type: array
                items:
                  $ref: '#/components/schemas/Dip'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  '/ddms/v2/dipsets/{dipsetid}/dips/query':
    get:
      tags:
        - Dips
      summary: Query dip from dipset
      description: Search dip within reference interval and specific classification
      operationId: query_dip
      parameters:
        - required: true
          schema:
            title: Dipsetid
            type: string
          name: dipsetid
          in: path
        - description: Min reference for the dips to search in the dipset
          required: false
          schema:
            title: Minreference
            type: number
            description: Min reference for the dips to search in the dipset
          name: minReference
          in: query
        - required: false
          schema:
            title: Max reference for the dips to search in the dipset
            type: number
          name: maxReference
          in: query
        - required: false
          schema:
            title: Classification for the dip to search in the dipset
            type: string
          name: classification
          in: query
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Query Dip Ddms V2 Dipsets  Dipsetid  Dips Query Get
                type: array
                items:
                  $ref: '#/components/schemas/Dip'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  '/ddms/v2/dipsets/{dipsetid}/dips/{index}':
    get:
      tags:
        - Dips
      summary: Get a dip at index
      description: '"Return dip from dipset at the given index'
      operationId: get_dip_by_index
      parameters:
        - required: true
          schema:
            title: Dipsetid
            type: string
          name: dipsetid
          in: path
        - required: true
          schema:
            title: Index
            type: integer
          name: index
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dip'
        '404':
          description: DipSet or index not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
    delete:
      tags:
        - Dips
      summary: Delete a dip
      description: Removes the dip at index
      operationId: delete_dip_by_index
      parameters:
        - required: true
          schema:
            title: Dipsetid
            type: string
          name: dipsetid
          in: path
        - required: true
          schema:
            title: Index
            type: integer
          name: index
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Delete Dip By Index Ddms V2 Dipsets  Dipsetid  Dips  Index  Delete
                type: array
                items:
                  $ref: '#/components/schemas/Dip'
        '404':
          description: DipSet or index not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
    patch:
      tags:
        - Dips
      summary: Update dip
      description: |-
        "Update dip at index
            Operation will sort by reference all dips in dipset (may modify dip indexes).
      operationId: patch_dip
      parameters:
        - required: true
          schema:
            title: Dipsetid
            type: string
          name: dipsetid
          in: path
        - required: true
          schema:
            title: Index
            type: integer
          name: index
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dip'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Patch Dip Ddms V2 Dipsets  Dipsetid  Dips  Index  Patch
                type: array
                items:
                  $ref: '#/components/schemas/Dip'
        '404':
          description: DipSet not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
...
components:
  schemas:
...
    Dip:
      title: Dip
      required:
        - reference
        - azimuth
        - inclination
      type: object
      properties:
        reference:
          title: Reference of the dip
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Only Measured Depth in meter is supported for the moment
        azimuth:
          title: Azimuth value of the dip
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Only degrees unit is supported for the moment
        inclination:
          title: Inclination value of the dip
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Only degrees unit is supported for the moment
        quality:
          title: Quality of the dip
          exclusiveMaximum: true
          maximum: 1
          exclusiveMinimum: false
          minimum: 0
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Decimal number between 0 and 1
        xCoordinate:
          title: The X coordinate of the dip
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Only meter unit is supported for the moment
        yCoordinate:
          title: The Y coordinate of the dip
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Only meter unit is supported for the moment
        zCoordinate:
          title: The Z coordinate of the dip
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Only meter unit is supported for the moment
        classification:
          title: Classification of the dip
          type: string
          description: Any string is accepted.
      example:
        reference:
          unitKey: meter
          value: 1000.5
        azimuth:
          unitKey: dega
          value: 42
        inclination:
          unitKey: dega
          value: 9
        quality:
          unitKey: unitless
          value: 0.5
        xCoordinate:
          unitKey: meter
          value: 2
        yCoordinate:
          unitKey: meter
          value: 45
        zCoordinate:
          unitKey: meter
          value: 7
        classification: fracture
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        errors:
          title: Errors
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    ValueWithUnit:
      title: ValueWithUnit
      required:
        - unitKey
        - value
      type: object
      properties:
        unitKey:
          title: Unit Key
          type: string
          description: Unit for value of the corresponding attribute for the domain object in question. The key can be looked up in the 'frameOfReference.units' for further details.
        value:
          title: Value
          type: number
          description: Value of the corresponding attribute for the domain object in question.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
  securitySchemes:
    OpenDESBearerToken:
      type: http
      scheme: bearer